changes done

+++ src/main/java/org/tweaktalent/hadoop/airline/AirLineMap.java
@@ -8,13 +8,13 @@ import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
-public class AirLineMap extends Mapper<IntWritable,Text,Text,IntWritable>{
+public class AirLineMap extends Mapper<LongWritable,Text,Text,IntWritable>{
 
 	private final static IntWritable one = new IntWritable(1);
     private Text word = new Text();
 	
-	//public void map(LongWritable key,Text value,Context context) throws IOException, InterruptedException {
-	public void map(IntWritable key,Text value,Context context) throws IOException, InterruptedException {
+	public void map(LongWritable key,Text value,Context context) throws IOException, InterruptedException {
+	//public void map(IntWritable key,Text value,Context context) throws IOException, InterruptedException {
 		
 		//String[] words = value.toString().split(",");
 		//if(words[0].equalsIgnoreCase("2008")) {
